# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "poran.io API Dokumentacija",
        "description": "Platforma za upravljanje sopstvenim vremenom",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Glavni API server"
        }
    ],
    "paths": {
...
        "/api/notes": {
            "get": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Prikaz svih beleški korisnika",
                "operationId": "9a95bde5931986ad60b9f7f414ec425e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Broj stranice za paginaciju",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beleške uspešno učitane",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Beleške uspešno učitane"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Note"
                                                    }
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Kreiranje nove beleške",
                "operationId": "d133c170557b4935898f2e444e0fd418",
                "requestBody": {
                    "description": "Podaci za kreiranje beleške",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Moja beleška"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Sadržaj beleške..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Beleška uspešno kreirana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Beleška uspešno kreirana"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Note"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Greška pri validaciji"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notes/{note}": {
            "get": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Prikaz određene beleške",
                "operationId": "e0f1599bca9af5cb49cb71f256ad31aa",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID beleške",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beleška uspešno učitanа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Note"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Beleška nije pronađena"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Ažuriranje beleške",
                "operationId": "a8fbf2764bddfa0b61a000b3d2b77721",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID beleške",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Podaci za ažuriranje beleške",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Ažurirana beleška"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Novi sadržaj beleške..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Beleška uspešno ažurirana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Beleška uspešno ažurirana"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Note"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Beleška nije pronađena"
                    },
                    "422": {
                        "description": "Greška pri validaciji"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Beleške"
                ],
                "summary": "Brisanje beleške",
                "operationId": "7f894d7b4f10aa080ffb3a1b2ebc5281",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID beleške",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beleška uspešno obrisana",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Beleška uspešno obrisana"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Beleška nije pronađena"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/notes": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Prikaz beleški određenog korisnika (Admin funkcija)",
                "operationId": "a56b7351e1b030b003a2835b61681ccb",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID korisnika",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beleške korisnika uspešno učitane",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Beleške korisnika Pera Peric uspešno učitane"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Note"
                                                    }
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nemate dozvolu za pristup"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Note": {
                "title": "Beleška",
                "description": "Model beleške u aplikaciji",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Moja beleška"
                    },
                    "content": {
                        "type": "string",
                        "example": "Sadržaj beleške..."
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "title": "Tag",
                "description": "Model taga za kategorizaciju beleški",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "važno"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "Korisnik",
                "description": "Model korisnika aplikacije",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pera Peric"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "pera@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "admin"
                        ],
                        "example": "user"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Unesite Bearer token dobijen pri prijavi",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}